ipq807x: add Q6v5 WCSS remoteproc support
Unlike previous ath10k based AHB SoC-s, ath11k is basically just an
interface to the dedicated Q6v5 DSP remote processor running the firmware
and radios.

So, add support for it based of various pending QCA patches.

IPQ8074 ath11k firmware is required for the remoteprocessor to work and
its part of every SoC, so include it by default for the target.

Signed-off-by: Robert Marko <robimarko@gmail.com>